backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-jose[cryptography]==3.3.0
supabase==2.9.1
stripe==10.11.0
Pillow==10.4.0
numpy==1.26.4
requests==2.32.3
python-dotenv==1.0.1
python-multipart==0.0.9
